+++ routes/Dashboard.js
@@ -1,12 +1,29 @@
 const express = require('express')
 const router = express.Router()
-const bcrypt = require('bcrypt')
 const bodyParser = require('body-parser')
+const db = require('../models')
+const bcrypt = require('bcrypt')
 const SALT = 15
 
-const db = require('../models')
-
 router.use(bodyParser.urlencoded({ extended: false }))
+
+router.post('/userInfo', (req, res) => {
+  const id = req.body.id
+
+  db.users
+    .findOne({
+      where: {
+        id: id
+      },
+      attributes: ['username', 'email']
+    })
+    .then((userInfo) => {
+      res.status(200).json(userInfo)
+    })
+    .catch((err) => {
+      console.log(err)
+    })
+})
 
 router.post('/updateUsername', async (req, res) => {
   const id = req.body.id
