--- routes/viewSavedGames.js
@@ -1,27 +1,0 @@
-const express = require('express')
-const router = express.Router()
-const bodyParser = require('body-parser')
-const db = require('../models')
-
-router.use(bodyParser.urlencoded({ extended: false }))
-
-router.post('/viewSavedGames', (req, res) => {
-  console.log(req.body)
-  const userID = req.body.userID
-  console.log(userID)
-
-  db.games
-    .findAll({
-      where: {
-        userID: userID
-      }
-    })
-    .then((games) => {
-      res.status(200).json(games)
-    })
-    .catch((err) => {
-      console.log(err)
-    })
-})
-
-module.exports = router
